requirementes.txt
colorama==0.3.7
decorator==4.0.10
Django==1.11.4
django-autocomplete-light==3.2.1
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-datatables-view==1.13.2
django-extensions==1.7.4
django-filter==1.0.4
django-tables2==1.10.0
djangorestframework==3.5.3
entrypoints==0.2.2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.0
lxml==3.6.4
MarkupSafe==0.23
mistune==0.7.3
mysqlclient==1.3.9
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.3
pandas==0.19.2
pickleshare==0.7.4
prompt-toolkit==1.0.8
psycopg2==2.6.2
Pygments==2.1.3
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.0
requests==2.12.4
simplegeneric==0.8.1
six==1.10.0
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
win-unicode-console==0.5
xlrd==1.0.0
